move gpu config to util function
