+++ 04_yolov3/model/model_factory.py
@@ -1,6 +1,7 @@
 import tensorflow as tf
 import model.backbone as back
 import model.head as head
+import utils.util_function as uf
 
 
 class ModelFactory:
@@ -47,20 +48,6 @@ def test_model_factory():
     print("!!! test_model_factory passed !!!")
 
 
-def set_gpu_configs():
-    gpus = tf.config.experimental.list_physical_devices('GPU')
-    if gpus:
-        try:
-            # Currently, memory growth needs to be the same across GPUs
-            for gpu in gpus:
-                tf.config.experimental.set_memory_growth(gpu, True)
-            logical_gpus = tf.config.experimental.list_logical_devices('GPU')
-            print(len(gpus), "Physical GPUs,", len(logical_gpus), "Logical GPUs")
-        except RuntimeError as e:
-            # Memory growth must be set before GPUs have been initialized
-            print(e)
-
-
 if __name__ == "__main__":
-    set_gpu_configs()
+    uf.set_gpu_configs()
     test_model_factory()
