+++ 04_yolov3/traineval/train_main.py
@@ -8,11 +8,12 @@ from config import Config as cfg
 from tfrecord.tfrecord_reader import TfrecordReader
 from model.model_factory import ModelFactory
 import traineval.train_val as tv
+import utils.util_function as uf
 import settings
 
 
 def train_main():
-    set_gpu_configs()
+    uf.set_gpu_configs()
     np.set_printoptions(precision=4, suppress=True, linewidth=100)
     for plan in cfg.Train.TRAINING_PLAN:
         train_by_plan(plan)
@@ -99,20 +100,6 @@ def read_previous_epoch():
     else:
         print(f"[read_previous_epoch] NO history in {filename}")
         return 0
-
-
-def set_gpu_configs():
-    gpus = tf.config.experimental.list_physical_devices('GPU')
-    if gpus:
-        try:
-            # Currently, memory growth needs to be the same across GPUs
-            for gpu in gpus:
-                tf.config.experimental.set_memory_growth(gpu, True)
-            logical_gpus = tf.config.experimental.list_logical_devices('GPU')
-            print(len(gpus), "Physical GPUs,", len(logical_gpus), "Logical GPUs")
-        except RuntimeError as e:
-            # Memory growth must be set before GPUs have been initialized
-            print(e)
 
 
 if __name__ == "__main__":
