+++ 04_yolov3/utils/util_function.py
@@ -2,6 +2,20 @@ import sys
 import tensorflow as tf
 
 
+def set_gpu_configs():
+    gpus = tf.config.experimental.list_physical_devices('GPU')
+    if gpus:
+        try:
+            # Currently, memory growth needs to be the same across GPUs
+            for gpu in gpus:
+                tf.config.experimental.set_memory_growth(gpu, True)
+            logical_gpus = tf.config.experimental.list_logical_devices('GPU')
+            print(len(gpus), "Physical GPUs,", len(logical_gpus), "Logical GPUs")
+        except RuntimeError as e:
+            # Memory growth must be set before GPUs have been initialized
+            print(e)
+
+
 def print_progress(status_msg):
     # Note the \r which means the line should overwrite itself.
     msg = "\r" + status_msg
